fix: adjust algolia index location
