+++ src/pages/searchWithFavorite/index.tsx
@@ -4,6 +4,15 @@ import { useHistory, useLocation } from "react-router-dom";
 import { useState } from "react";
 import { IArticle } from "../../hooks/useAlgoArticles";
 import { TabList, Tab, TabPanel, TabPanels } from "../../components";
+import algoliasearch from "algoliasearch";
+
+const client = algoliasearch(
+  process.env.REACT_APP_ALGOLIA_APLICATION_ID as string,
+  process.env.REACT_APP_ALGOLIA_APLICATION_KEY as string
+);
+const index = client.initIndex(
+  process.env.REACT_APP_ALGOLIA_INDEX_NAME as string
+);
 
 const Index = () => {
   const history = useHistory();
@@ -51,7 +60,11 @@ const Index = () => {
       </TabList>
       <TabPanels>
         <TabPanel isCur={location.pathname === "/search"}>
-          <Search handleButtonClick={handleButtonClick} favorites={favorites} />
+          <Search
+            handleButtonClick={handleButtonClick}
+            favorites={favorites}
+            index={index}
+          />
         </TabPanel>
         <TabPanel isCur={location.pathname === "/favorite"}>
           <Favorite
